requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
numpy>=1.21,<2.1
torch --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
python-jose[cryptography]
supabase-py
python-magic
moviepy